Relocation of Tests - now in logical parent package
